Improve parse_dates interface
So far the inferring the parse_dates input argument depended on two
values in the schema. The dtype of a column being datetime.date
populated the parse_dates list, and combining two columns into a
datetime column depended on another schema parameter called
"combine_dt_cols". All that is scrapped, everything rests on the
"parse_dates" parameter in the schema now.